network-visualizer/backend/requirements.txt
flask==2.0.1
werkzeug==2.0.3
pyyaml==6.0.2
omegaconf==2.1.1
flask-cors==3.0.10
requests==2.28.1
